Original change: https://android-review.googlesource.com/c/platform/packages/modules/Bluetooth/+/2259842

+++ system/gd/crypto_toolbox/aes_cmac.cc
@@ -100,7 +100,7 @@ static Octet16 cmac_aes_k_calculate(const Octet16& key) {
   Octet16 output;
   Octet16 x{0};  // zero initialized
 
-  uint8_t i = 1;
+  uint16_t i = 1;
   while (i <= cmac_cb.round) {
     /* Mi' := Mi (+) X  */
     xor_128((Octet16*)&cmac_cb.text[(cmac_cb.round - i) * OCTET16_LEN], x);
